Move parse_working_time_range to timr_utils
