+++ tests/test_timr_utils.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock, patch, MagicMock, call
 import datetime
-from timr_utils import UIProjectTime, ProjectTimeConsolidator
+from timr_utils import UIProjectTime, ProjectTimeConsolidator, parse_working_time_range
 
 
 class TestUIProjectTime(unittest.TestCase):
@@ -735,6 +735,30 @@ class TestProjectTimeConsolidatorEdgeCases(unittest.TestCase):
         # Verify the result matches mock return value
         self.assertEqual(result, mock_result)
 
+
+class TestWorkingTimeUtils(unittest.TestCase):
+    """Tests for working time helper functions"""
+
+    def test_parse_range_uses_duration(self):
+        working_time = {
+            "start": "2025-06-14T10:00:00+00:00",
+            "end": None,
+            "duration": {"minutes": 15},
+        }
+        start, end = parse_working_time_range(working_time)
+        self.assertEqual(
+            start,
+            datetime.datetime(2025, 6, 14, 10, 0, tzinfo=datetime.timezone.utc),
+        )
+        self.assertEqual(end - start, datetime.timedelta(minutes=15))
+
+    def test_parse_range_missing_values(self):
+        with self.assertRaises(ValueError):
+            parse_working_time_range({"start": "2025-06-14T10:00:00+00:00"})
+
+    def test_parse_range_missing_start(self):
+        with self.assertRaises(ValueError):
+            parse_working_time_range({"end": "2025-06-14T10:10:00+00:00"})
 
 if __name__ == '__main__':
     unittest.main()
